+++ app/controllers/recording_studio_terms_and_conditions/application_controller.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
 module RecordingStudioTermsAndConditions
-  class ApplicationController < ActionController::Base
+  class ApplicationController < (defined?(::ApplicationController) ? ::ApplicationController : ActionController::Base)
+    include RecordingStudio::UsesDefaultLayout if defined?(RecordingStudio::UsesDefaultLayout)
+    helper RecordingStudioTermsAndConditions::ApplicationHelper
+
     protect_from_forgery with: :exception
-    layout "application"
   end
 end