+++ test/dummy/config/initializers/recording_studio.rb
@@ -3,6 +3,7 @@
 RecordingStudio.configure do |config|
   # Registered delegated_type recordables (strings or classes)
   config.recordable_types = [
+    "AdminRoot",
     "Workspace",
     "Folder",
     "Page",